BloscPlugin: hack/work-around issue where Frame data type is not defined
If the data type is not defined then we cannot determine the size of the datatype and an exception is thrown.
We work around this by hard-coding the datatype size to 2 bytes as this is what the Excalibur detector output.

See JIRA https://jira.diamond.ac.uk/browse/BC-816 and https://jira.diamond.ac.uk/browse/BC-811